login y registro funciona

+++ routes/web.php
@@ -10,28 +10,23 @@
 | contains the "web" middleware group. Now create something great!
 |
 */
-//Route::post('/productos/agregar',  );
-Route::delete('admin/productos/eliminar/{id}', 'ProductController@deleteProduct');
-Route::post('admin/productos/editar/{id}', 'ProductController@updateProduct');
-Route::get('admin/productos/editar/{id}', 'ProductController@editProduct');
-Route::get('admin/productos', 'ProductController@createProduct');
-Route::post('admin/productos', 'ProductController@createProductSave');
-//Route::get('/home', 'HomeController@index')->name('home');
+
+Route::delete('admin/productos/eliminar/{id}', 'ProductController@deleteProduct')->middleware('auth');
+Route::post('admin/productos/editar/{id}', 'ProductController@updateProduct')->middleware('auth');
+Route::get('admin/productos/editar/{id}', 'ProductController@editProduct')->middleware('auth');
+Route::get('admin/productos', 'ProductController@createProduct')->middleware('auth')->middleware('auth');
+Route::post('admin/productos', 'ProductController@createProductSave')->middleware('auth');
+
 
 Route::get('/shop/{product_id}','ProductController@showProduct');
-// Route::get('/shop/{category_name}/{category_id}/{page}','ProductController@showCategory');
+
 Route::get('/shop/{category_name}/{category_id}','ProductController@showCategory');
 Route::get('/shop','ProductController@showShop');
-//Route::get('/productAdmin','ProductController@showProductAdmin');
+
 Route::get('/contacto','ContactController@showContact');
 Route::get('/ayuda','HelpController@showHelp');
 Route::get('/carrito','CartController@showCart');
-Route::get('/registro',function(){
-  return view('register');
-});
-Route::get('/login',function(){
-  return view('login');
-});
+
 Route::get('/', 'IndexController@showIndex');
 
 
@@ -39,4 +34,8 @@ Route::get('/', 'IndexController@showIndex');
 
 
 
+
+
 Auth::routes();
+
+Route::get('/home', 'HomeController@index')->name('home');
